+++ CHANGELOG.md
@@ -23,6 +23,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed issue with test-runner fancy UI where it would crash if the summary was too large.
 
+### Removed
+- Removed support for the deprecated `maelstrom-test.toml` file.
+
 ## [0.11.0] - 2024-07-30
 
 ### General